requirements/common.txt
Connexion[swagger-ui]==2.7.0
Flask-Sqlalchemy==2.4.4
Flask-Login==0.5.0
Python-dotenv==0.15.0
Pymysql==0.10.1
Python-codicefiscale==0.3.7
Faker==4.16.0
pybreaker==0.6.0 